+++ spec/dummy_rails_app/app/models/pizza.rb
@@ -1,5 +1,6 @@
 class Pizza < ApplicationRecord
   belongs_to :programmer, dependent: :delete_recursively # lest he starve
+  belongs_to :baker, class_name: 'Pizza::Baker', dependent: :destroy
   has_one :box, dependent: :delete_recursively
   has_many :toppings, class_name: 'Ingredient',
                       foreign_key: :dish_id,