+++ spec/dummy_rails_app/app/models/pizza/baker.rb
@@ -1,0 +1,4 @@
+class Pizza < ApplicationRecord
+  class Baker < ApplicationRecord
+  end
+end