+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+
+## v1.1.0
 ### Added
 - support for polymorphic associations
 - `ActiveRecord::Base#delete_recursively`